need to solve vtk memory leak problem, details about vtkSmartPointer

+++ includes/robotModel.h
@@ -10,18 +10,18 @@ public://methods
     RobotModel();
     ~RobotModel();
     
-    void updateDHs(float* DHS);
-    std::vector<float> getThetas();
-    void setup (float* DHs);
+    void updateDHs(double* DHS);
+    std::vector<double> getThetas();
+    void setup (double* DHs);
     void update();
 
     vtkSmartPointer<vtkTransform> getEndEffectorTransform();
     vtkSmartPointer<vtkMatrix4x4> getEndEffectorMatrix();
-    void calcInverseKinematics(const Eigen::Matrix< float , 6 , 1>& goal);
+    void calcInverseKinematics(const Eigen::Matrix< double , 6 , 1>& goal);
 private://methods
-	Eigen::Matrix4f calcHomoTransMatrix(float d, float a, float alpha, float theta);
-    Eigen::Matrix3f calcRotTransMatrix(float d, float a, float alpha, float theta);
-    Eigen::MatrixXf calcJacobian();
+	Eigen::Matrix4d calcHomoTransMatrix(double d, double a, double alpha, double theta);
+    Eigen::Matrix3d calcRotTransMatrix(double d, double a, double alpha, double theta);
+    Eigen::MatrixXd calcJacobian();
 
 public://members
 	
@@ -30,9 +30,9 @@ private://members
     vtkSmartPointer<vtkMatrix4x4> _endEffectorMatrix;
 	GraphicalModel** _joints;
 	GraphicalModel** _links;
-    float** _dh_parameters;
-    Eigen::Matrix< float , 6 , 1> _twistVector;
-    float _lastDeltaPos;
+    double** _dh_parameters;
+    Eigen::Matrix< double , 6 , 1> _twistVector;
+    double _lastDeltaPos;
 protected: 
     
 };
